+++ src/components/Rider/RiderProfile.jsx
@@ -1,5 +1,5 @@
 import { FontAwesomeIcon as Icon } from "@fortawesome/react-fontawesome";
-import { faCircleUser, faSearch } from "@fortawesome/free-solid-svg-icons";
+import { faCircleUser } from "@fortawesome/free-solid-svg-icons";
 import { Col, Row, Input } from "reactstrap";
 const RiderProfile = () => {
   return (
@@ -13,7 +13,7 @@ const RiderProfile = () => {
 
         <div className="rect-profile">
           <Row>
-            <Col md={3}>
+            <Col md={2}>
               <Icon
                 className="profile-img"
                 icon={faCircleUser}
@@ -33,7 +33,8 @@ const RiderProfile = () => {
                   Age:20 &nbsp; &nbsp;&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;
                   &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;
                   &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;
-                  &nbsp; &nbsp; &nbsp; Vehicle No: 47548
+                  &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; Vehicle No:
+                  47548
                   <br></br>
                   Contact Number: 09564457889
                   <br></br>
@@ -41,10 +42,6 @@ const RiderProfile = () => {
               </Row>
             </Col>
           </Row>
-        </div>
-
-        <div className="search-box">
-          <input type="text" placeholder="Search for rider" />
         </div>
       </div>
     </>
