+++ src/routes/RiderPage.jsx
@@ -1,6 +1,7 @@
 import "../assets/css/RiderPage.css";
 import RiderTable from "../components/Rider/RiderTable.jsx";
 import RiderProfile from "../components/Rider/RiderProfile.jsx";
+import { Row } from "reactstrap";
 const RiderPage = () => {
   return (
     <>
